Split users table to reflect database changes
As discussed previously the database tables have changed.
The users table has been replaced with the usernames and
usercontacts tables. The reason for the split is the need
to separate user sensitive data into the usercontacts table.
Aside from that, this commit is about macthing the changes in
the database to their structs in the acct file.

Changes made to accts file

- removed the users struct

- added usernames table
  - has the fields
    - id
    - username

- added usercontacts table
  - has the fields
    - id
    - user_id
    - fname
    - lname
    - address